+++ R/mts_filterDate.R
@@ -1,9 +1,9 @@
 #' @export
 #' @importFrom rlang .data
 #'
-#' @title Date filtering for MazamaSingleTimeseries objects
+#' @title Date filtering for \emph{mts} objects
 #'
-#' @param sts MazamaSingleTimeseries \emph{sts} object.
+#' @param mts \emph{mts} object.
 #' @param startdate Desired start datetime (ISO 8601).
 #' @param enddate Desired end datetime (ISO 8601).
 #' @param timezone Olson timezone used to interpret dates.
@@ -15,9 +15,9 @@
 #'   \code{\link[lubridate]{ceiling_date}} to the \code{enddate} rather than
 #'   \code{\link[lubridate]{floor_date}}
 #'
-#' @description Subsets a MazamaSingleTimeseries object by date. This function
+#' @description Subsets a \emph{mts} object by date. This function
 #' always filters to day-boundaries. For sub-day filtering, use
-#' \code{sts_filterDatetime()}.
+#' \code{mts_filterDatetime()}.
 #'
 #' Dates can be anything that is understood by \code{MazamaCoreUtils::parseDatetime()}
 #' including either of the following recommended formats:
@@ -33,7 +33,7 @@
 #' \enumerate{
 #' \item{get timezone from \code{startdate} if it is \code{POSIXct}}
 #' \item{use passed in \code{timezone}}
-#' \item{get timezone from \code{sts}}
+#' \item{get timezone from \code{mts}}
 #' }
 #'
 #' @note The returned data will run from the beginning of \code{startdate} until
@@ -42,21 +42,21 @@
 #' \code{enddate} is less than 24 hours after \code{startdate}. In that case, a
 #' single day is returned.
 #'
-#' @return A subset of the given \emph{sts} object.
+#' @return A subset of the given \emph{mts} object.
 #'
-#' @seealso \link{sts_filter}
-#' @seealso \link{sts_filterDatetime}
+#' @seealso \link{mts_filter}
+#' @seealso \link{mts_filterDatetime}
 #' @examples
 #' library(MazamaTimeSeries)
 #'
-#' example_sts %>%
-#'   sts_filterDate(startdate = 20180808, enddate = 20180815) %>%
-#'   sts_extractData() %>%
+#' example_mts %>%
+#'   mts_filterDate(startdate = 20180808, enddate = 20180815) %>%
+#'   mts_extractData() %>%
 #'   head()
 #'
 
-sts_filterDate <- function(
-  sts = NULL,
+mts_filterDate <- function(
+  mts = NULL,
   startdate = NULL,
   enddate = NULL,
   timezone = NULL,
@@ -67,16 +67,25 @@ sts_filterDate <- function(
 
   # ----- Validate parameters --------------------------------------------------
 
-  MazamaCoreUtils::stopIfNull(sts)
+  # A little involved to catch the case where the user forgets to pass in 'mts'
 
-  if ( !sts_isValid(sts) )
-    stop("Parameter 'sts' is not a valid 'sts' object.")
+  result <- try({
+    if ( !mts_isValid(mts) )
+      stop("First argument is not a valid 'mts' object.")
+  }, silent = TRUE)
 
-  if ( sts_isEmpty(sts) )
-    stop("Parameter 'sts' has no data.")
+  if ( class(result) %in% "try-error" ) {
+    err_msg <- geterrmessage()
+    if ( stringr::str_detect(err_msg, "object .* not found") ) {
+      stop(paste0(err_msg, "\n(Did you forget to pass in the 'mts' object?)"))
+    }
+  }
+
+  if ( mts_isEmpty(mts) )
+    stop("Parameter 'mts' has no data.")
 
   # Remove any duplicate data records
-  sts <- sts_distinct(sts)
+  mts <- mts_distinct(mts)
 
   if ( is.null(startdate) && !is.null(enddate) )
     stop("At least one of 'startdate' or 'enddate' must be specified")
@@ -85,7 +94,7 @@ sts_filterDate <- function(
   # POSIXct times then you know what you are doing.
   #   1) get timezone from startdate if it is POSIXct
   #   2) use passed in timezone
-  #   3) get timezone from sts
+  #   3) get timezone from mts
 
   if ( lubridate::is.POSIXt(startdate) ) {
 
@@ -97,12 +106,12 @@ sts_filterDate <- function(
 
   } else {
 
-    # Handle multiple timezones in 'sts'
-    timezoneCount <- length(unique(sts$meta$timezone))
+    # Handle multiple timezones in 'mts'
+    timezoneCount <- length(unique(mts$meta$timezone))
 
     # Use table(timezone) to find the most common one
     if ( timezoneCount > 1 ) {
-      timezoneTable <- sort(table(sts$meta$timezone), decreasing = TRUE)
+      timezoneTable <- sort(table(mts$meta$timezone), decreasing = TRUE)
       timezone <- names(timezoneTable)[1]
       warning(sprintf(
         "Found %d timezones. Only %s will be used.",
@@ -110,7 +119,7 @@ sts_filterDate <- function(
         timezone
       ))
     } else {
-      timezone <- sts$meta$timezone[1]
+      timezone <- mts$meta$timezone[1]
     }
 
   }
@@ -129,35 +138,35 @@ sts_filterDate <- function(
     ceilingEnd = ceilingEnd,
   )
 
-  # ----- Subset the 'sts' object ----------------------------------------------
+  # ----- Subset the 'mts' object ----------------------------------------------
 
   # NOTE:  When processing lots of data automatically, it is best not to stop()
   # NOTE:  when no data exist for a requested date range. Instead, return a
-  # NOTE:  valid 'sts' object with zero rows of data.
+  # NOTE:  valid 'mts' object with zero rows of data.
 
-  if (dateRange[1] > sts$data$datetime[length(sts$data$datetime)] |
-      dateRange[2] < sts$data$datetime[1]) {
+  if (dateRange[1] > mts$data$datetime[length(mts$data$datetime)] |
+      dateRange[2] < mts$data$datetime[1]) {
 
     message(sprintf(
-      "sts (%s) does not contain requested date range",
-      sts$meta$siteName
+      "mts (%s) does not contain requested date range",
+      mts$meta$siteName
     ))
 
-    data <- sts$data[0,]
+    data <- mts$data[0,]
 
   } else {
 
     data <-
-    sts$data %>%
-    dplyr::filter(.data$datetime >= dateRange[1]) %>%
-    dplyr::filter(.data$datetime < dateRange[2])
+      mts$data %>%
+      dplyr::filter(.data$datetime >= dateRange[1]) %>%
+      dplyr::filter(.data$datetime < dateRange[2])
 
   }
 
-  sts$data <- data
+  mts$data <- data
 
   # ----- Return ---------------------------------------------------------------
 
-  return(sts)
+  return(mts)
 
 }