+++ R/mts_filterDatetime.R
@@ -1,9 +1,9 @@
 #' @export
 #' @importFrom rlang .data
 #'
-#' @title Datetime filtering for MazamaSingleTimeseries objects
+#' @title Datetime filtering for \code{mts} objects
 #'
-#' @param sts MazamaSingleTimeseries \emph{sts} object.
+#' @param mts \emph{mts} object.
 #' @param startdate Desired start datetime (ISO 8601).
 #' @param enddate Desired end datetime (ISO 8601).
 #' @param timezone Olson timezone used to interpret dates.
@@ -15,8 +15,8 @@
 #'   \code{\link[lubridate]{ceiling_date}} to the \code{enddate} rather than
 #'   \code{\link[lubridate]{floor_date}}
 #'
-#' @description Subsets a MazamaSingleTimeseries object by datetime. This function
-#' allows for sub-day filtering as opposed to \code{sts_filterDate()} which
+#' @description Subsets a \code{mts} object by datetime. This function
+#' allows for sub-day filtering as opposed to \code{mts_filterDate()} which
 #' always filters to day-boundaries.
 #'
 #' Datetimes can be anything that is understood by
@@ -29,28 +29,28 @@
 #' \enumerate{
 #' \item{get timezone from \code{startdate} if it is \code{POSIXct}}
 #' \item{use passed in \code{timezone}}
-#' \item{get timezone from \code{sts}}
+#' \item{get timezone from \code{mts}}
 #' }
 #'
-#' @return A subset of the given \emph{sts} object.
+#' @return A subset of the given \emph{mts} object.
 #'
-#' @seealso \link{sts_filter}
-#' @seealso \link{sts_filterDate}
+#' @seealso \link{mts_filter}
+#' @seealso \link{mts_filterDate}
 #'
 #' @examples
 #' library(MazamaTimeSeries)
 #'
-#' example_sts %>%
-#'   sts_filterDatetime(
+#' example_mts %>%
+#'   mts_filterDatetime(
 #'     startdate = "2018-08-08 06:00:00",
 #'     enddate = "2018-08-14 18:00:00"
 #'   ) %>%
-#'   sts_extractData() %>%
+#'   mts_extractData() %>%
 #'   head()
 #'
 
-sts_filterDatetime <- function(
-  sts = NULL,
+mts_filterDatetime <- function(
+  mts = NULL,
   startdate = NULL,
   enddate = NULL,
   timezone = NULL,
@@ -61,24 +61,24 @@ sts_filterDatetime <- function(
 
   # ----- Validate parameters --------------------------------------------------
 
-  MazamaCoreUtils::stopIfNull(sts)
+  MazamaCoreUtils::stopIfNull(mts)
   MazamaCoreUtils::stopIfNull(startdate)
   MazamaCoreUtils::stopIfNull(enddate)
 
-  if ( !sts_isValid(sts) )
-    stop("Parameter 'sts' is not a valid 'sts' object.")
+  if ( !mts_isValid(mts) )
+    stop("Parameter 'mts' is not a valid 'mts' object.")
 
-  if ( sts_isEmpty(sts) )
-    stop("Parameter 'sts' has no data.")
+  if ( mts_isEmpty(mts) )
+    stop("Parameter 'mts' has no data.")
 
   # Remove any duplicate data records
-  sts <- sts_distinct(sts)
+  mts <- mts_distinct(mts)
 
   # Timezone determination precedence assumes that if you are passing in
   # POSIXct times then you know what you are doing.
   #   1) get timezone from startdate if it is POSIXct
   #   2) use passed in timezone
-  #   3) get timezone from sts
+  #   3) get timezone from mts
 
   if ( lubridate::is.POSIXt(startdate) ) {
 
@@ -90,12 +90,12 @@ sts_filterDatetime <- function(
 
   } else {
 
-    # Handle multiple timezones in 'sts'
-    timezoneCount <- length(unique(sts$meta$timezone))
+    # Handle multiple timezones in 'mts'
+    timezoneCount <- length(unique(mts$meta$timezone))
 
     # Use table(timezone) to find the most common one
     if ( timezoneCount > 1 ) {
-      timezoneTable <- sort(table(sts$meta$timezone), decreasing = TRUE)
+      timezoneTable <- sort(table(mts$meta$timezone), decreasing = TRUE)
       timezone <- names(timezoneTable)[1]
       warning(sprintf(
         "Found %d timezones. Only %s will be used.",
@@ -103,7 +103,7 @@ sts_filterDatetime <- function(
         timezone
       ))
     } else {
-      timezone <- sts$meta$timezone[1]
+      timezone <- mts$meta$timezone[1]
     }
 
   }
@@ -122,35 +122,35 @@ sts_filterDatetime <- function(
     ceilingEnd = ceilingEnd
   )
 
-  # ----- Subset the 'sts' object ----------------------------------------------
+  # ----- Subset the 'mts' object ----------------------------------------------
 
   # NOTE:  When processing lots of data automatically, it is best not to stop()
   # NOTE:  when no data exist for a requested date range. Instead, return a
-  # NOTE:  valid 'sts' object with zero rows of data.
+  # NOTE:  valid 'mts' object with zero rows of data.
 
-  if (timeRange[1] > sts$data$datetime[length(sts$data$datetime)] |
-      timeRange[2] < sts$data$datetime[1]) {
+  if (timeRange[1] > mts$data$datetime[length(mts$data$datetime)] |
+      timeRange[2] < mts$data$datetime[1]) {
 
     message(sprintf(
-      "sts (%s) does not contain requested time range",
-      sts$meta$siteName
+      "mts (%s) does not contain requested time range",
+      mts$meta$siteName
     ))
 
-    data <- sts$data[0,]
+    data <- mts$data[0,]
 
   } else {
 
     data <-
-      sts$data %>%
+      mts$data %>%
       dplyr::filter(.data$datetime >= timeRange[1]) %>%
       dplyr::filter(.data$datetime < timeRange[2])
 
   }
 
-  sts$data <- data
+  mts$data <- data
 
   # ----- Return ---------------------------------------------------------------
 
-  return(sts)
+  return(mts)
 
 }